First fileImporter.py
pyForm GUI used.
The user chooses all the files they wish to import. Only the GUI of the file that needs to be imported is enabled at any time. 
The user needs to click on the "Import" button in order to move on to importing the next file - this is temporary.
File types are detected by checking the file extensions - this is temporary.
The "Done" button is only enabled once all data files have been imported.
The option to import a fam file is only given if the user imports a Admix data file.
The "redo importing" is a temporary fix to an issue I couldn't solve due to not being able to check when the user clicks on a GUI's respective "open" button. I had to manually loop through checking each input so the "redo" button just resets the form (in case they accidentally import the wrong file).
Clicking the import button's function returns a dictionary containing the fileName, the fileType and the fileData.